+++ backend/src/api/conference-room-reservation/services/conference-room-reservation.js
@@ -1,0 +1,9 @@
+'use strict';
+
+/**
+ * conference-room-reservation service.
+ */
+
+const { createCoreService } = require('@strapi/strapi').factories;
+
+module.exports = createCoreService('api::conference-room-reservation.conference-room-reservation');