+++ backend/src/api/conference-room/controllers/conference-room.js
@@ -1,0 +1,9 @@
+'use strict';
+
+/**
+ *  conference-room controller
+ */
+
+const { createCoreController } = require('@strapi/strapi').factories;
+
+module.exports = createCoreController('api::conference-room.conference-room');