+++ backend/src/api/conference-room/routes/conference-room.js
@@ -1,0 +1,9 @@
+'use strict';
+
+/**
+ * conference-room router.
+ */
+
+const { createCoreRouter } = require('@strapi/strapi').factories;
+
+module.exports = createCoreRouter('api::conference-room.conference-room');