+++ backend/src/api/desk-blockage/controllers/desk-blockage.js
@@ -1,0 +1,9 @@
+'use strict';
+
+/**
+ *  desk-blockage controller
+ */
+
+const { createCoreController } = require('@strapi/strapi').factories;
+
+module.exports = createCoreController('api::desk-blockage.desk-blockage');