+++ backend/src/api/desk-blockage/routes/desk-blockage.js
@@ -1,0 +1,9 @@
+'use strict';
+
+/**
+ * desk-blockage router.
+ */
+
+const { createCoreRouter } = require('@strapi/strapi').factories;
+
+module.exports = createCoreRouter('api::desk-blockage.desk-blockage');