+++ backend/src/api/desk-cancellation/routes/desk-cancellation.js
@@ -1,0 +1,9 @@
+'use strict';
+
+/**
+ * desk-cancellation router.
+ */
+
+const { createCoreRouter } = require('@strapi/strapi').factories;
+
+module.exports = createCoreRouter('api::desk-cancellation.desk-cancellation');