+++ backend/src/api/desk/controllers/desk.js
@@ -1,0 +1,9 @@
+'use strict';
+
+/**
+ *  desk controller
+ */
+
+const { createCoreController } = require('@strapi/strapi').factories;
+
+module.exports = createCoreController('api::desk.desk');